requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
python-dotenv==1.0.1
Werkzeug==3.0.2
